Implement ma_unitdatax_request

+++ MAC_SAP/fyp/r_n_g.h
@@ -3,6 +3,6 @@
 
 #include "common.h"
 
-double random_int_gen(uint32 *state);
+float random_int_gen(uint32 *state);
 
 #endif
